backend/requirements.txt
Django==4.0.6
djangorestframework==3.12.4
djoser==2.1.0
django-colorfield
django-filter
drf_extra_fields
